NIFI-5102 Modified License and Notice files; Changed package names
