+++ nifi-nar-bundles/nifi-marklogic-bundle/nifi-marklogic-processors/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -13,5 +13,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-com.marklogic.nifi.processor.PutMarkLogic
-com.marklogic.nifi.processor.QueryMarkLogic
+org.apache.nifi.marklogic.processor.PutMarkLogic
+org.apache.nifi.marklogic.processor.QueryMarkLogic
