+++ nifi-nar-bundles/nifi-marklogic-bundle/nifi-marklogic-services/src/main/resources/META-INF/services/org.apache.nifi.controller.ControllerService
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-com.marklogic.nifi.controller.DefaultMarkLogicDatabaseClientService
+org.apache.nifi.marklogic.controller.DefaultMarkLogicDatabaseClientService
